requirements.txt
Flask==2.2.2
Flask-Cors==4.0.0
python-dotenv==1.0.1
Jinja2==3.1.3
Werkzeug==2.2.2
matplotlib==3.8.2
plotly==5.18.0
scikit-learn==1.4.0
scipy==1.12.0
numpy==1.26.4
pandas==2.2.0
typing_extensions==4.13.2
tenacity==8.2.3
openai==1.82.1
neo4j
gunicorn